Fixed issue 216: Adding custom properties with same name results in CrashSender.exe assertion failure.
git-svn-id: http://crashrpt.googlecode.com/svn/trunk@1636 821a4644-a21a-0410-aec7-6199d47568cc

+++ processing/crprober/main.cpp
@@ -351,7 +351,7 @@ int process_report(LPTSTR szInput, LPTSTR szInputMD5, LPTSTR szOutput,
             int pos = sMD5DirName.rfind('\\');
             if(pos<0) // There is no back slash in path
                 sMD5DirName = _T(""); 
-            else if(pos!=sMD5DirName.length()-1) // Append the back slash to dir name
+            else if(pos!=(int)sMD5DirName.length()-1) // Append the back slash to dir name
                 sMD5DirName = sMD5DirName.substr(0, pos+1);
         }
     }
